+++ src/test/groovy/ru/endlesscode/bukkitgradle/server/extension/ServerConfigurationSpec.groovy
@@ -2,13 +2,13 @@ package ru.endlesscode.bukkitgradle.server.extension
 
 import spock.lang.Specification
 
-class RunConfigurationSpec extends Specification {
+class ServerConfigurationSpec extends Specification {
 
     // SUT
-    RunConfiguration runConfiguration
+    ServerConfigurationImpl runConfiguration
 
     void setup() {
-        runConfiguration = new RunConfiguration()
+        runConfiguration = new ServerConfigurationImpl()
     }
 
     void 'when build args - should return args with debug flags'() {
@@ -25,14 +25,6 @@ class RunConfigurationSpec extends Specification {
 
         when:
         def args = runConfiguration.buildJvmArgs()
-
-        then:
-        "-Dfile.encoding=UTF-8 -Xmx1G" == args
-    }
-
-    void 'when build args with debug mode override - should return args without debug flags'() {
-        when:
-        def args = runConfiguration.buildJvmArgs(false)
 
         then:
         "-Dfile.encoding=UTF-8 -Xmx1G" == args